+++ push_swap/srcs/ft_sort.c
@@ -16,24 +16,11 @@
 void	ft_sort_three(t_list **list)
 {
 	if (ft_sorted(*list))
-		return;
+		return ;
 	if (ft_list_min(*list) == (*list)->content)
-	{
-		ft_swap_head(list);
-		ft_printf("sa\n");
-		ft_rotate_up(list);
-		ft_printf("ra\n");
-	}
+		ft_case_min_at_top(list);
 	else if (ft_list_max(*list) == (*list)->content)
-	{
-		ft_rotate_up(list);
-		ft_printf("ra\n");
-		if (ft_sorted(*list) == 0)
-		{
-			ft_swap_head(list);
-			ft_printf("sa\n");
-		}
-	}
+		ft_case_max_at_top(list);
 	else
 	{
 		if ((*list)->content > ((*list)->next)->content)
@@ -46,6 +33,27 @@ void	ft_sort_three(t_list **list)
 			ft_rotate_down(list);
 			ft_printf("rra\n");
 		}
+	}
+}
+
+//Cuando el minimo esta arriba
+void	ft_case_min_at_top(t_list **list)
+{
+	ft_swap_head(list);
+	ft_printf("sa\n");
+	ft_rotate_up(list);
+	ft_printf("ra\n");
+}
+
+//Cuando el maximo esta arriba
+void	ft_case_max_at_top(t_list **list)
+{
+	ft_rotate_up(list);
+	ft_printf("ra\n");
+	if (ft_sorted(*list) == 0)
+	{
+		ft_swap_head(list);
+		ft_printf("sa\n");
 	}
 }
