+++ push_swap/srcs/mechanical_turk.c
@@ -19,12 +19,12 @@ void	ft_mechanical_turk(t_list **stack_A, t_list **stack_B)
 
 	i = 0;
 	if (ft_sorted(*stack_A))
-		return;
+		return ;
 	if (ft_list_len(*stack_A) == 2 && ft_sorted(*stack_A) == 0)
 	{
 		ft_swap_head(stack_A);
 		ft_printf("sa\n");
-		return;
+		return ;
 	}
 	while (ft_list_len(*stack_A) > 3 && i++ < 2)
 	{
@@ -36,17 +36,10 @@ void	ft_mechanical_turk(t_list **stack_A, t_list **stack_B)
 	if (ft_list_len(*stack_A) == 3)
 		ft_sort_three(stack_A);
 	if (*stack_B != NULL && ft_list_len(*stack_A) == 3)
-	{
-		//ft_highest_at_top(stack_B);
 		ft_push_b(stack_A, stack_B);
-	}
 	if (ft_sorted(*stack_A) == 0)
 		ft_lowest_at_top(stack_A);
 }
-	/*
-	faltaria ver si es necesario rotar A hasta que el minimo este en el top
-	y comprobar si la lista esta ordenada
-	*/
 
 //Pushea los numeros de A a B de forma ordenada
 void	ft_sort_b(t_list **stack_A, t_list **stack_B)
@@ -59,91 +52,11 @@ void	ft_sort_b(t_list **stack_A, t_list **stack_B)
 	len = ft_list_len(*stack_A);
 	while (i < len - 3)
 	{
-		
-		//printf("STACK A\tSTACK B\n");
-		//ft_print_2list(*stack_A,*stack_B);
-		//getchar();
 		fastest_number = ft_fastest_number(*stack_A, *stack_B);
-		//getchar();
 		ft_sort_num(stack_A, stack_B, fastest_number, 'd');
-		//getchar();
 		ft_push_between(stack_A, stack_B);
 		ft_printf("pb\n");
 		i++;
-		//getchar();
-	}
-}
-
-//Coloca el numero mas alto de B en el top
-void	ft_highest_at_top(t_list **list)
-{
-	int	max;
-	int	moves_up;
-	int	moves_down;
-
-	max = ft_list_max(*list);
-	moves_up = ft_moves_up(*list, max);
-	moves_down = ft_moves_down(*list, max);
-	while ((*list)->content != max)
-	{
-		if (moves_up < moves_down)
-		{
-			ft_rotate_up(list);
-			ft_printf("rb\n");
-		}
-		else
-		{
-			ft_rotate_down(list);
-			ft_printf("rrb\n");
-		}
-	}
-}
-
-//Coloca el numero mas bajo de A en el top
-void	ft_lowest_at_top(t_list **list)
-{
-	int	min;
-	int	moves_up;
-	int	moves_down;
-
-	min = ft_list_min(*list);
-	moves_up = ft_moves_up(*list, min);
-	moves_down = ft_moves_down(*list, min);
-	while ((*list)->content != min)
-	{
-		if (moves_up < moves_down)
-		{
-			ft_rotate_up(list);
-			ft_printf("ra\n");
-		}
-		else
-		{
-			ft_rotate_down(list);
-			ft_printf("rra\n");
-		}
-	}
-}
-
-//Coloca el numero mas bajo de A en el top
-void	ft_num_at_top(t_list **list, int number)
-{
-	int	moves_up;
-	int	moves_down;
-
-	moves_up = ft_moves_up(*list, number);
-	moves_down = ft_moves_down(*list, number);
-	while ((*list)->content != number)
-	{
-		if (moves_up < moves_down)
-		{
-			ft_rotate_up(list);
-			ft_printf("ra\n");
-		}
-		else
-		{
-			ft_rotate_down(list);
-			ft_printf("rra\n");
-		}
 	}
 }
 
@@ -153,7 +66,7 @@ void	ft_push_b(t_list **stack_A, t_list **stack_B)
 {
 	int	i;
 	int	len;
-	int place;
+	int	place;
 
 	i = 0;
 	len = ft_list_len(*stack_B);
@@ -167,15 +80,6 @@ void	ft_push_b(t_list **stack_A, t_list **stack_B)
 			i++;
 		}
 		else
-		{
-			//ft_sort_num(stack_B, stack_A, (*stack_B)->content, 'a');
-			//printf("Stack A:\n");
-			//ft_print_list(*stack_A);
-			
 			ft_num_at_top(stack_A, place);
-
-			//ft_rotate_down(stack_A);
-			//ft_printf("rra\n");
-		}
 	}
 }
